Register

+++ base/views.py
@@ -36,6 +36,21 @@ def loginPage(request):
     context = {'page': page}
     return render(request, 'base/login_register.html', context)
 
+def registerPage(request):
+    form = MyUserCreationForm()
+
+    if request.method == 'POST':
+        form = MyUserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error occurred during registration')
+
+    return render(request, 'base/login_register.html', {'form': form})
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -63,34 +78,55 @@ def room(request,pk):
     return render(request,'base/room.html',context)
 
 
+@login_required(login_url='login')
 def createRoom(request):
-    form=RoomForm()
+    form = RoomForm()
+    topics = Topic.objects.all()
+    if request.method == 'POST':
+        topic_name = request.POST.get('topic')
+        topic, created = Topic.objects.get_or_create(name=topic_name)
 
-    if request.method =='POST':
-        form=RoomForm(request.POST)
-        if form.is_valid():
-            form.save()
-            return redirect('home')
-
-    context={'form':form}
-    return render(request,'base/room_form.html',context)
-
-def updateRoom(request,pk):
-    room=Room.objects.get(id=pk)
-    form=RoomForm(instance=room)
-    if request.method=='POST':
-        form=RoomForm(request.POST,instance=room)
-        if form.is_valid():
-            form.save()
-            return redirect('home')
-    context={'form':form}
-    return render(request,'base/room_form.html',context)
-
-def deleteRoom(request,pk):
-    room=Room.objects.get(id=pk)
-    if request.method=='POST':
-        room.delete() 
+        Room.objects.create(
+            host=request.user,
+            topic=topic,
+            name=request.POST.get('name'),
+            description=request.POST.get('description'),
+        )
         return redirect('home')
-        
-    return render(request,'base/delete.html',{'obj':room})
 
+    context = {'form': form, 'topics': topics}
+    return render(request, 'base/room_form.html', context)
+
+
+@login_required(login_url='login')
+def updateRoom(request, pk):
+    room = Room.objects.get(id=pk)
+    form = RoomForm(instance=room)
+    topics = Topic.objects.all()
+    if request.user != room.host:
+        return HttpResponse('Your are not allowed here!!')
+
+    if request.method == 'POST':
+        topic_name = request.POST.get('topic')
+        topic, created = Topic.objects.get_or_create(name=topic_name)
+        room.name = request.POST.get('name')
+        room.topic = topic
+        room.description = request.POST.get('description')
+        room.save()
+        return redirect('home')
+
+    context = {'form': form, 'topics': topics, 'room': room}
+    return render(request, 'base/room_form.html', context)
+
+
+@login_required(login_url='login')
+def deleteRoom(request, pk):
+    room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('Your are not allowed here!!')
+
+    if request.method == 'POST':
+        room.delete()
+        return redirect('home')
+    return render(request, 'base/delete.html', {'obj': room})
